Switch to using focusPanel store
